Begin to address the swtich statement code smell
This interface isn't open for new requirements to the types of readers for the players. To get to an open state we need to refactor by addressing a code smell -> revaluate if the code is open -> if not, then pick a new code smell and start the process again.

In this case it's quite easy to see the swtch statement code smell. Since we can identify it, we know that there is a recipe to refactor it. In this case we can use the "Replace Type Code with State/Strategy" recipe to remove the logic of the read types to classes, using composition.

Note: There is another recipe to handle this, " Replace Type Code with Subclasses " however, this approach uses inheritance, in this case, since a reader should "read" and this is a behaviour, composition makes more sense to me.