Create a factory for the PlayerViews
The final step in this refactor recipe is moving the selection of the Player view class out of PlayersObject to the newly created factory. This now opens the PlayersObject up to use new display methods.

To add a new display simply add a class for the new display logic (that extends the PlayerView interface), then add it to the factory.